requirements.txt
dateparser
Pillow
scrapyd
scrapy-deltafetch
scrapy-fake-useragent
scrapy-jsonschema
scrapy-magicfields
scrapy-splash
scrapyd-client
ScrapyElasticSearch
usaddress
